requirements-prods.txt
# Build/runtime: gunicorn still imports pkg_resources, so keep setuptools below 81
setuptools<81

# Core Flask Framework
Flask==2.2.5
gunicorn==21.2.0
Werkzeug==2.2.3

# Database
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
psycopg2-binary>=2.9.6

# Authentication & Security
flask-jwt-extended==4.4.4
flask-bcrypt==1.0.1
Flask-Limiter==3.5.0
Flask-WTF==1.2.1

# Background Tasks & Caching
redis==4.6.0
rq==1.1.0

# Cloud Storage
cloudinary==1.29.0
boto3==1.26.99

# Report Generation
docxtpl>=0.16.0
openpyxl==3.1.2
XlsxWriter==3.1.2
reportlab==4.4.6
pandas>=2.3.3
xlrd>=2.0.1
# pandas.read_html() for CAFM "Excel" exports that are actually HTML tables
beautifulsoup4>=4.12.0
lxml>=5.0.0
html5lib>=1.1
Pillow>=11.0.0
docx2pdf>=0.1.8
python-docx>=1.0.0

# Utilities
requests==2.31.0
python-dotenv==1.0.0
tenacity==8.2.3
marshmallow==3.20.1

# Scheduler (MMR daily email)
APScheduler>=3.10.0